feat(fe): add page useeffect
