+++ apps/frontend/components/Analytics.tsx
@@ -1,0 +1,52 @@
+'use client'
+
+import { usePathname, useSearchParams } from 'next/navigation'
+import Script from 'next/script'
+import { useEffect } from 'react'
+
+declare global {
+  interface Window {
+    gtag: (
+      command: 'config' | 'event',
+      action: string,
+      options?: {
+        [key: string]: string | number | boolean | undefined
+      }
+    ) => void
+  }
+}
+
+export function Analytics() {
+  const pathname = usePathname()
+  const searchParams = useSearchParams()
+
+  useEffect(() => {
+    if (pathname) {
+      const url = pathname + searchParams.toString()
+      window.gtag('event', 'page_view', {
+        page_path: url
+      })
+    }
+  }, [pathname, searchParams])
+
+  return (
+    <>
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}`}
+      />
+      <Script
+        id="gtag-init"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}');
+          `
+        }}
+      />
+    </>
+  )
+}
